% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createWKT.r
\name{createWKT}
\alias{createWKT}
\title{Create polygon from point}
\usage{
createWKT(latitude, longitude, radius = 5000)
}
\arguments{
\item{latitude}{numeric}

\item{longitude}{numeric}

\item{radius}{numeric, distance in meters}
}
\value{
Polygon as a WKT string
}
\description{
Covert latitude and longitude plus a radius into a WKT string.
}
\examples{
createWKT(51.6011023, -1.1278673)
}

