% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGroups.R
\name{listGroups}
\alias{listGroups}
\title{Get group definitions}
\usage{
listGroups(...)
}
\arguments{
\item{...}{Further named parameters passed on to \code{\link[httr]{GET}}}
}
\value{
A dataframe containing the definitions of groups on the NBN Gateway
}
\description{
Gives a dataframe of the group definitions from the NBN Gateway for reference.
}
\examples{
\dontrun{ 
 t <- listGroups()
}
}
\author{
Stuart Ball, JNCC \email{stuart.ball@jncc.gov.uk}
}
\seealso{
\code{\link{getFeature}}, \code{\link{getOccurrences}}
}

