% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVCs.R
\name{listVCs}
\alias{listVCs}
\title{Get list of vice-counties}
\usage{
listVCs()
}
\value{
A dataframe containing the name and featureID of each vice-county.
}
\description{
Gives a dataframe of the Watsonian vice-counties and their keys for reference.
}
\examples{
\dontrun{ 
 t <- listVCs()
}
}
\author{
Stuart Ball, JNCC \email{stuart.ball@jncc.gov.uk}
}
\seealso{
\code{\link{getFeature}}, \code{\link{getOccurrences}}
}

