\docType{package}
\name{rnbn-package}
\alias{rnbn}
\alias{rnbn-package}
\title{Access data from the NBN}
\description{
  Access data from the NBN
}
\details{
  The National Buiodiversity Network (NBN) is a repository
  for biodiversity data in the UK. Access is provided via
  web-services. At the time of writing, RESTful
  web-services are under test and are used here, but the
  URLs can be expected to change once a production version
  is released. This package provides R functions to access
  data via these web-services.
}
\author{
  Stuart Ball, JNCC \email{stuart.ball@jncc.gov.uk}
}
\references{
  \url{http://http://data.nbn.org.uk/Documentation/Web_Services/}
}
\keyword{package}

