\docType{package}
\name{rnbn-package}
\alias{rnbn}
\title{Access data from the NBN}
\description{
Access data from the NBN
}
\details{
The National Biodiversity Network (NBN) is a repository for biodiversity data in the UK. Access is provided via web-services. This package provides R functions to access data via these web-services. For a tutorial see the vignette included with the package
}
\author{
Stuart Ball, JNCC \email{stuart.ball@jncc.gov.uk} and Tom August, CEH \email{tomaug@ceh.ac.uk}
}
\references{
\url{http://http://data.nbn.org.uk/Documentation/Web_Services/}
}
\keyword{package}

