\name{listGroups}
\alias{listGroups}
\title{Get group definitions}
\usage{
  listGroups()
}
\value{
  A dataframe containing the definitions of groups on the
  NBN Gateway
}
\description{
  Gives a dataframe of the group definitions from the NBN
  Gateway for reference.
}
\examples{
\dontrun{
 t <- listGroups()
}
}
\author{
  Stuart Ball, JNCC \email{stuart.ball@jncc.gov.uk}
}
\seealso{
  \code{\link{getFeature}}, \code{\link{getOccurrences}}
}

